Created 2 cases to sign up a student user and a service staff user.
In this commit, I created 2 cases to sign up a student user and a service staff user.  There are different sign up experiences for students and service staff (the system requests the user to input a set of credentials for students which are different from the input feilds requested of a service staff).